+++ include/tvm/ir/span.h
@@ -79,22 +79,23 @@ class Span;
  */
 class SpanNode : public Object {
  public:
-  /*! \brief The source name */
+  /*! \brief The source name. */
   SourceName source;
-  /*! \brief Line number */
-  int lineno;
-  /*! \brief column offset */
-  int col_offset;
+  /*! \brief The line number. */
+  int line;
+  /*! \brief The column offset. */
+  int column;
+
   // override attr visitor
   void VisitAttrs(AttrVisitor* v) {
     v->Visit("source", &source);
-    v->Visit("lineno", &lineno);
-    v->Visit("col_offset", &col_offset);
+    v->Visit("line", &line);
+    v->Visit("column", &column);
   }
 
   bool SEqualReduce(const SpanNode* other, SEqualReducer equal) const {
-    return equal(source, other->source) && equal(lineno, other->lineno) &&
-           equal(col_offset, other->col_offset);
+    return equal(source, other->source) && equal(line, other->line) &&
+           equal(column, other->column);
   }
 
   static constexpr const char* _type_key = "Span";
